% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_stat_text.R
\name{highlight_stat_text}
\alias{highlight_stat_text}
\title{Highlight statistical text}
\usage{
highlight_stat_text(text, file)
}
\arguments{
\item{text}{text to highlight}

\item{file}{file name}
}
\value{
creates a .html file
}
\description{
Highlight statistical text
}
