% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html.R
\name{make_html}
\alias{make_html}
\title{Make html}
\usage{
make_html(count_stats_tbl)
}
\arguments{
\item{count_stats_tbl}{Tibble generated using the count_stats() function.}
}
\value{
Text string of html
}
\description{
Make html
}
