% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html_file.R
\name{make_html_file}
\alias{make_html_file}
\title{Make html file}
\usage{
make_html_file(html_text, file)
}
\arguments{
\item{html_text}{HTML text}

\item{file}{file name}
}
\description{
Make html file
}
