% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html_text.R
\name{make_html_text}
\alias{make_html_text}
\title{Make html text}
\usage{
make_html_text(count_stats_tbl)
}
\arguments{
\item{count_stats_tbl}{Tibble generated using the count_stats() function.}
}
\value{
Text string of html
}
\description{
Make html text
}
